backend/pyproject.toml
[project]
name = "fast-prisma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "prisma>=0.15.0",
    "colorama>=0.4.6",
    "types-colorama>=0.4.15.20240311",
    "faker>=33.1.0",
    "tokenizer>=3.4.5",
    "zstandard>=0.23.0",
    "vibrato>=0.2.0",
    "rapidfuzz>=3.11.0",
    "python-jose>=3.3.0",
    "types-python-jose>=3.3.4.20240106",
    "requests>=2.32.3",
    "types-requests>=2.32.0.20241016",
    "cachetools>=5.5.0",
    "types-cachetools>=5.5.0.20240820",
    "pyjwt>=2.10.1",
    "cryptography>=44.0.0",
]
[base_fastapi]
ignore_missing_imports = true
[tool.setuptools]
packages = ["app", "log", "docker"]
